Support conversation-style notifications
In the previous commit, we moved to using MessagingStyle notifications which offloaded a lot of work that we were doing in formatting to the framework. Now, we can benefit from conversation style notifications with the addition of linking our notifications to shortcuts and bubbles.

This additionally adds a mark as read notification action